Changing a DataType Of all interfaces

+++ src/main/java/tn/esprit/foyeresprit/Service/IBlocService.java
@@ -5,7 +5,7 @@ import tn.esprit.foyeresprit.Entity.Bloc;
 import java.util.List;
 
 public interface IBlocService {
-    Bloc getBlocById(int id);
+    Bloc getBlocById(Long id);
     Bloc addBloc(Bloc bloc);
     Bloc updateBloc(Bloc bloc);
     void deleteBloc(Bloc bloc);
