+++ src/main/java/tn/esprit/foyeresprit/Service/IEtudiantService.java
@@ -5,9 +5,9 @@ import tn.esprit.foyeresprit.Entity.Etudiant;
 import java.util.List;
 
 public interface IEtudiantService {
-    Etudiant getEtudiant(int id);
+    Etudiant getEtudiant(Long id);
     List<Etudiant> getEtudiants();
     Etudiant addEtudiant(Etudiant et);
     Etudiant updateEtudiant(Etudiant et);
-    void deleteEtudiant(int id);
+    void deleteEtudiant(Long id);
 }
