+++ src/main/java/tn/esprit/foyeresprit/Service/IReservationService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IReservationService {
     Reservation addReservation(Reservation reservation);
     List<Reservation> getAllReservations();
-    Reservation getReservationById(int id);
+    Reservation getReservationById(Long id);
     Reservation updateReservation(Reservation reservation);
-    void deleteReservation(int id);
+    void deleteReservation(Long id);
 }
